pyproject.toml
[tool.poetry]
name = "clv-analytics"
version = "0.1.0"
description = "CLV Analytics System Backend"
authors = ["Matthew <mattbarg@unc.edu>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
alembic = "^1.17.2"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.1.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

